test: move packages test to Packages.test.tsx file
this commit move relevant packages test to packages.test.tsx
